% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db.R
\name{check_db}
\alias{check_db}
\title{Sicherheitscheck für die \code{db_data.rds} einer Universität}
\usage{
check_db(test_data)
}
\arguments{
\item{test_data}{Ein Data Frame mit den zu prüfenden Daten.}
}
\value{
Ein einzelnes \code{ptblank_agent}-Objekt mit den Prüfergebnissen
Zusätzlich wird ein HTML-Report im Viewer angezeigt.
}
\description{
\code{check_db} prüft mithilfe des Paketes \code{pointblank} verschiedene Struktur-, Typ-
und Plausibilitätsprüfungen für \code{db_data.rds}. Der Check bildet den Abschluss des Cleaningprozesses einer Universität.
}
\details{
Folgende Prüfungen werden durchgeführt:
\enumerate{
\item \strong{Vollständigkeits-Check:} Alle erwarteten Variablen (z. B. \code{hochschule},
\code{jahr}, \code{semester}, Future-Skills-Variablen etc.) sind im Datensatz enthalten.
\item \strong{Datentypen-Check Strings:} Alle inhaltlichen Text-Variablen (außer den
Future-Skills-Variablen und der Hilfsspalte \code{kursbeschreibung_len}) müssen
vom Typ \code{character} sein.
\item \strong{Datentypen-Check Numeric:} Die Future-Skills-Variablen \code{data_analytics_ki},
\code{softwareentwicklung}, \code{nutzerzentriertes_design}, \code{it_architektur},
\code{hardware_robotikentwicklung} und \code{quantencomputing} müssen numerisch sein.
\item \strong{Mindestlänge Kursbeschreibung:} Die Länge der
Kursbeschreibung (\code{kursbeschreibung_len}) muss mindestens 20 Zeichen betragen.
Kürzere Beschreibungen gelten als fehlerhaft und sollten NA gesetzt werden.
\item \strong{Hochschulnamen:} Die Werte in \code{hochschule} müssen in der geladenen
Referenzliste zulässiger Hochschulnamen enthalten sein (\code{inst/extdata/hochschulen_namen_kuerzel.sql}).
\item \strong{Hochschulkürzel:} Die Werte in \code{hochschule_kurz} müssen in der entsprechenden
Referenzliste zulässiger Kürzel enthalten sein (\code{inst/extdata/hochschulen_namen_kuerzel.sql}).
\item \strong{Pflichtfelder Jahr/Semester:} Die Felder \code{jahr} und \code{semester} dürfen
keine fehlenden Werte (\code{NA}) enthalten.
\item \strong{Sprach-Codierung:} Die Werte in \code{sprache_recoded} müssen zu der im Wiki definierten, erlaubten
Menge gehören, z. B. \code{"Deutsch"}, \code{"Englisch"}, \code{"Deutsch/Englisch"},
weitere Sprachen, \code{"Sonstiges"} oder \code{NA}.
\item \strong{Kursformat-Codierung:} Die Werte in \code{kursformat_recoded} müssen zu der im Wiki definierten,
festen Menge gehören, z. B. \code{"Vorlesung"}, \code{"Seminar"}, \code{"Übung"}, \code{"Austausch"},
\code{"Erfahrung"}, \code{"Sprachkurs"}, \code{"Sonstiges"} oder \code{NA}.
\item \strong{Semester-Format:} Die Werte in \code{semester} müssen dem Muster \code{"YYYYs"} oder
\code{"YYYYw"} entsprechen (vierstellige Jahreszahl, gefolgt von \code{s} für Sommer-
bzw. \code{w} für Wintersemester).
}
}
