% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nas.R
\name{check_nas}
\alias{check_nas}
\title{Visualisiert NA-Konzentration nach Semester}
\usage{
check_nas(data, semester)
}
\arguments{
\item{data}{Ein data.frame oder tibble mit den zu analysierenden Daten.}

\item{semester}{Unquoted Spaltenname mit Semester-Informationen (z.B. \verb{WS 2019/20} oder \code{2019}).}
}
\value{
Ein \code{ggplot}-Objekt mit der Visualisierung der NA-Konzentration.
}
\description{
Erzeugt einen Plot mit der NA-Konzentration pro Variable und sortiert die
Semester-Faktoren nach der darin enthaltenen vierstelligen Jahreszahl. Falls
keine vierstellige Jahreszahl in der angegebenen Spalte gefunden wird,
bricht die Funktion mit einer aussagekräftigen Fehlermeldung ab.
}
