% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook_and_db_data.R
\name{create_codebook_and_db_data}
\alias{create_codebook_and_db_data}
\title{Erstellt \code{codebook} und \code{db_data} aus Rohdaten}
\usage{
create_codebook_and_db_data(raw_data, raw_data_fs = NULL)
}
\arguments{
\item{raw_data}{Data Frame mit den Rohdaten.}

\item{raw_data_fs}{Data Frame mit Future-Skills-Spalten.}
}
\value{
Eine Liste mit zwei Elementen:
\describe{
\item{codebook}{Ein Tibble mit einer Spalte \code{Variablen}.}
\item{db_data}{Ein Tibble im erwarteten HEX-DB-Format.}
}
}
\description{
Diese Funktion bildet einen typischen Schritt im HEX-Cleaning ab: Aus
\code{raw_data} wird ein \code{codebook} mit den vorhandenen Variablennamen erzeugt und
daraus anschliessend ein \code{db_data}-Datensatz im erwarteten HEX-Format
aufgebaut. Die benoetigten Spalten in \code{raw_data} und \code{raw_data_fs} muessen
vollstaendig vorhanden sein. Nur die fachlichen Metadatenfelder, die erst
spaeter im Prozess befuellt werden, werden mit \code{NA} angelegt.
}
\examples{
raw_data <- tibble::tibble(
  anmerkungen = "Hinweis",
  dozierende = "Prof. Beispiel",
  ects = "5",
  fakultaet = "Informatik",
  hochschule = "Universitaet Musterstadt",
  hochschule_kurz = "UMS",
  jahr = "2026",
  kursbeschreibung = "Dies ist eine ausreichend lange Kursbeschreibung.",
  kursformat_original = "Vorlesung",
  kursformat_recoded = "Vorlesung",
  lehrtyp = "Pflicht",
  lernmethode = "Praesenz",
  lernziele = "Lernziel",
  literatur = "Buch",
  module = "Modul A",
  nummer = "101",
  organisation_orig = "Original",
  organisation = "Organisation",
  pfad = "/tmp/pfad",
  pruefung = "Klausur",
  scrape_datum = "2026-04-01",
  semester = "2026s",
  sprache_original = "Deutsch",
  sprache_recoded = "Deutsch",
  studiengaenge = "BSc",
  sws = "2",
  teilnehmerzahl = "30",
  titel = "Kurs A",
  url = "https://example.org",
  voraussetzungen = "Keine",
  zusatzinformationen = "Info",
  institut = "Institut A"
)

raw_data_fs <- tibble::tibble(
  data_analytics_ki = 1,
  softwareentwicklung = 0,
  nutzerzentriertes_design = 0,
  it_architektur = 1,
  hardware_robotikentwicklung = 0,
  quantencomputing = 0
)

res <- create_codebook_and_db_data(raw_data, raw_data_fs)
res$codebook
res$db_data

}
