% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_str_to_na.R
\name{empty_str_to_na}
\alias{empty_str_to_na}
\title{Ersetzt leere Strings durch \code{NA}}
\usage{
empty_str_to_na(df)
}
\arguments{
\item{df}{Ein \code{data.frame} oder \code{tibble}.}
}
\value{
Ein Objekt mit derselben Struktur wie \code{df}, in dem leere oder nur aus
Leerzeichen bestehende Zeichenketten in Character-Spalten durch \code{NA}
ersetzt wurden.

empty_str_to_na(test_df)
}
\description{
Bereinigt alle Zeichenketten-Spalten eines Data Frames, indem führende und
nachgestellte Leerzeichen entfernt und anschließend leere Strings (\code{""}) als
fehlende Werte (\code{NA}) gesetzt werden.
}
