% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_rds.R
\name{json_to_rds}
\alias{json_to_rds}
\title{Konvertiert eine JSON-Datei in ein RDS-Format}
\usage{
json_to_rds(input_path, output_path)
}
\arguments{
\item{input_path}{Pfad zur Eingabe-JSON-Datei.}

\item{output_path}{Pfad zur Ausgabedatei im RDS-Format.}
}
\value{
Das eingelesene Tibble.
}
\description{
Diese Funktion liest eine JSON-Datei ein, wandelt sie in ein Tibble um und speichert das Ergebnis als RDS-Datei.
}
