% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_semantic_na_values.R
\name{remove_semantic_na_values}
\alias{remove_semantic_na_values}
\title{Setzt Werte einer String-Variable NA, wenn diese weniger als 20 Zeichen enthalten}
\usage{
remove_semantic_na_values(texts, min_num_letters = 20)
}
\arguments{
\item{texts}{Ein Vektor von Zeichenketten (Character-Vektor), der bereinigt werden soll.}

\item{min_num_letters}{Minimale Anzahl an Buchstaben (Standard: 20), die ein Text enthalten muss,
damit er nicht als semantisches NA betrachtet wird.}
}
\value{
Ein Character-Vektor, in dem zu kurze Texte durch \code{NA} ersetzt wurden.
}
\description{
Diese Funktion ersetzt Einträge in einem Vektor von Texten durch \code{NA},
wenn die Anzahl der Buchstaben (ohne Satzzeichen) kleiner als ein definierter Schwellenwert ist.
}
\examples{
\dontrun{
library(dplyr)   
db_data_universitaet_jena  |>
mutate(
kursbeschreibung_cleaned = remove_semantic_na_values(kursbeschreibung, min_num_letters = 30)
)
}
}
