% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_character_columns.R
\name{squish_character_columns}
\alias{squish_character_columns}
\title{Fuehrt \code{str_squish()} auf allen Character-Variablen aus}
\usage{
squish_character_columns(data)
}
\arguments{
\item{data}{Ein \code{data.frame} oder \code{tibble}.}
}
\value{
Ein Datensatz, in dem alle Character-Variablen mit
\code{stringr::str_squish()} bereinigt wurden.
}
\description{
Bereinigt alle Character-Spalten eines Datensatzes, indem fuehrende und
nachgestellte Leerzeichen entfernt und Mehrfach-Leerzeichen innerhalb von
Strings auf ein einzelnes Leerzeichen reduziert werden. Die Funktion ist fuer
Pipelines gedacht und gibt den bereinigten Datensatz zurueck.
}
\examples{
\dontrun{
final_df |>
  squish_character_columns()
}

}
